+++ src/activities/Summary.js
@@ -94,7 +94,7 @@ require(['jquery', 'i18next', 'ysdtemplate',
                        contentType : 'application/json; charset=utf-8',
                        crossDomain: true,
                        success: function(data, textStatus, jqXHR) {
-
+                         /*
                          if (model.requestLanguage != data.customer_language &&
                              data.customer_language != null &&
                              data.customer_language != '') {
@@ -102,9 +102,10 @@ require(['jquery', 'i18next', 'ysdtemplate',
                                                   commonServices.orderUrl + '?id=' + data.free_access_id;
                          }
                          else {
+                         */ 
                            model.order = data;
                            view.updateOrder();
-                         }
+                         //}
 
                        },
                        error: function(data, textStatus, jqXHR) {